requirements.txt
pandas>=1.5.3
gtts>=2.3.1
playsound>=1.3.0
numpy>=1.23.5
pydub>=0.25.1
scipy>=1.10.1
vosk>=0.3.45
openai>=0.27.4
openai-whisper>=20230314
pyaudio>=0.2.13